final design of GameplaySongbar
expose Expanded for animation trigger, add shadow effect

custom team color
