+++ osu.Game.Tournament/Screens/Gameplay/GameplayScreen.cs
@@ -140,8 +140,6 @@ namespace osu.Game.Tournament.Screens.Gameplay
                 warmupButton.Alpha = !w.NewValue ? 0.5f : 1;
                 header.ShowScores = !w.NewValue;
             }, true);
-
-            SongBar.Margin = new MarginPadding { Bottom = 7f };
         }
 
         protected override void LoadComplete()
